app/services/vox_scribe/requirements.txt
pyannote.audio
faster-whisper
qdrant-client
torch
torchaudio
python-dotenv
pydub
noisereduce
soundfile
speechbrain